adds dockerfile
