+++ requirements.txt
@@ -1,4 +1,1 @@
-jupyter==1.0.0
-pandas==1.2.0
-pymc3==3.10.0
-seaborn==0.11.1
+pystan
